// frontend/react_dash/src/redux/types.ts
export interface SalesData {
    year: number;
    month_no: number;
    month_name?: string;
    sales_amount: number;
    net_profit: number;
    customer_name: string;
    quantity: number;
    region: string;
    sigungu: string;
    product_name: string;
    product_category_name: string;
    category_name: string;
}